Split db-restore common.ts into utils modules
- utils/psql.ts — quoting and psql query/exec/check primitives
- utils/env.ts — .env loading and DbUrls
- utils/restoreSteps.ts — the restore pipeline steps
- utils/runScript.ts — CLI error harness
